requirements.txt
cryptography
pyotp
qrcode
Pillow
pyperclip
pynput
